+++ src/widgets/animimage/lv_animimage.c
@@ -71,7 +71,7 @@ void lv_animimg_set_src(lv_obj_t * obj, const void * dsc[], size_t num)
     LV_ASSERT_OBJ(obj, MY_CLASS);
     lv_animimg_t * animimg = (lv_animimg_t *)obj;
     animimg->dsc = dsc;
-    animimg->pic_count = num;
+    animimg->pic_count = (int8_t)num;
     lv_anim_set_values(&animimg->anim, 0, (int32_t)num - 1);
 }
